software: remove `AccessArguments`. NFCI
