+++ software/glasgow/access/direct/__init__.py
@@ -1,6 +1,5 @@
-from .arguments import DirectArguments
 from .multiplexer import DirectMultiplexer
 from .demultiplexer import DirectDemultiplexer
 
 
-__all__ = ["DirectArguments", "DirectMultiplexer", "DirectDemultiplexer"]
+__all__ = ["DirectMultiplexer", "DirectDemultiplexer"]
