+++ software/glasgow/access/simulation/__init__.py
@@ -1,6 +1,5 @@
-from .arguments import *
 from .multiplexer import *
 from .demultiplexer import *
 
 
-__all__ = ["SimulationArguments", "SimulationMultiplexer", "SimulationDemultiplexer"]
+__all__ = ["SimulationMultiplexer", "SimulationDemultiplexer"]
